# openapi.yaml
openapi: 3.0.0
info:
  title: API E-commerce Sederhana
  description: API untuk mengelola pengguna, toko, produk, pesanan, dan transaksi.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Server Lokal
components:
  schemas:
...
    UserRole:
      type: string
      enum: [buyer, seller, admin]
      description: Peran pengguna.
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
        full_name:
          type: string
        email:
          type: string
          format: email
        password_hash:
          type: string
          description: Hash kata sandi (Hanya ada di database).
          readOnly: true
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
          readOnly: true
    NewUser:
      type: object
      required: [email, password, full_name]
      properties:
        full_name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          description: Kata sandi mentah.
          example: S3cur3P@ssw0rd!
        phone_number:
          type: string
          example: "081234567890"
...
paths:
...
  /users/register:
    post:
      tags:
        - Users
      summary: Mendaftarkan Pengguna Baru
      description: Membuat akun pengguna baru dengan peran default 'buyer'.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '201':
          description: Pengguna berhasil dibuat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email sudah terdaftar.
  /users/{userId}:
    get: # ✅ GET DITAMBAHKAN
      tags:
        - Users
      summary: Mendapatkan Detail Pengguna
      description: Mengambil detail profil pengguna berdasarkan ID.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID pengguna.
      responses:
        '200':
          description: Detail pengguna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Pengguna tidak ditemukan.
    put:
      tags:
        - Users
      summary: Memperbarui Profil Pengguna
      description: Memperbarui detail profil pengguna yang ada.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: ID pengguna yang akan diperbarui.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
              properties:
                full_name:
                  type: string
                phone_number:
                  type: string
      responses:
        '200':
          description: Profil berhasil diperbarui.
        '404':
          description: Pengguna tidak ditemukan.
    delete: # ✅ DELETE DITAMBAHKAN
      tags:
        - Users
      summary: Menghapus Akun Pengguna
      description: Menghapus akun pengguna. (Hanya oleh Admin atau pengguna itu sendiri)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Akun berhasil dihapus (No Content).
        '404':
          description: Pengguna tidak ditemukan.

  # -----------------------------------------------------
  # --- ENDPOINT PRODUK (products) ---
  # -----------------------------------------------------
...